Split out the population of settings and refdata, so refdata is in its own file
